requirements.txt
psycopg2-binary>=2.9
Faker>=18.0.0
python-dotenv>=1.0.0
